fix: update yextVisualEditorPlugin (#369)
Updated main.tsx template and confirmed dev mode and production builds
work

+++ packages/visual-editor/src/vite-plugin/plugin.ts
@@ -19,6 +19,9 @@ export const yextVisualEditorPlugin = (): Plugin => {
       const filePath = path.join(process.cwd(), fileName);
       filesToCleanup.push(filePath);
       fs.mkdirSync(path.dirname(filePath), { recursive: true });
+      if (fs.existsSync(filePath)) {
+        return;
+      }
       fs.writeFileSync(filePath, content);
     });
   };
